requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
chromadb
langchain
langchain_google_genai
faiss-cpu
langchain-community
groq
langchain-groq
pypdf
streamlit_option_menu
